Make flag parameters unsigned longs

+++ include/linted/mq.h
@@ -82,7 +82,8 @@ struct linted_mq_task_send
  * @error ENOSPC Insufficient space.
  */
 linted_error linted_mq_create(linted_mq * restrict mqp, char const *debugpath,
-                              struct linted_mq_attr const *attr, int flags);
+                              struct linted_mq_attr const *attr,
+                              unsigned long flags);
 
 void linted_mq_task_receive(struct linted_mq_task_receive *task,
                             unsigned task_action, linted_ko ko, char *buf,
