+++ include/linted/updater.h
@@ -85,7 +85,8 @@ struct linted_updater_task_receive
 
 static linted_updater_int linted_updater__saturate(int_fast64_t x);
 
-linted_error linted_updater_create(linted_updater * restrict updaterp, int flags);
+linted_error linted_updater_create(linted_updater * restrict updaterp,
+                                   unsigned long flags);
 
 void linted_updater_send(struct linted_updater_task_send *task, int task_id,
                          linted_updater updater,
