+++ src/ko/ko.c
@@ -74,7 +74,7 @@ linted_error linted_ko_dummy(linted_ko * restrict kop)
 }
 
 linted_error linted_ko_open(linted_ko * restrict kop, linted_ko dirko,
-                            char const *pathname, int flags)
+                            char const *pathname, unsigned long flags)
 {
     linted_error errnum;
 
@@ -143,7 +143,7 @@ linted_error linted_ko_open(linted_ko * restrict kop, linted_ko dirko,
 }
 
 linted_error linted_ko_reopen(linted_ko * restrict kooutp,
-                              linted_ko koin, int flags)
+                              linted_ko koin, unsigned long flags)
 {
     char pathname[sizeof "/proc/self/fd/" + 10U];
     sprintf(pathname, "/proc/self/fd/%i", koin);
